+++ src/lib.rs
@@ -55,10 +55,11 @@ pub use crate::tables::gdef::{
 pub use crate::tables::gpos::GposTable as GposView;
 pub use crate::tables::gsub::GsubTable as GsubView;
 pub use crate::tables::gsub::{
-    Ligature, LigatureComponentIter, LigatureSet, LigatureSubst, LigatureSubstIter, SingleSubst,
-    SingleSubstIter, GSUB_LOOKUP_TYPE_ALTERNATE, GSUB_LOOKUP_TYPE_CHAINED_CONTEXT,
-    GSUB_LOOKUP_TYPE_CONTEXT, GSUB_LOOKUP_TYPE_EXTENSION, GSUB_LOOKUP_TYPE_LIGATURE,
-    GSUB_LOOKUP_TYPE_MULTIPLE, GSUB_LOOKUP_TYPE_REVERSE_CHAINED_SINGLE, GSUB_LOOKUP_TYPE_SINGLE,
+    Ligature, LigatureComponentIter, LigatureSet, LigatureSubst, LigatureSubstIter, MultipleSubst,
+    MultipleSubstIter, Sequence, SequenceGlyphIter, SingleSubst, SingleSubstIter,
+    GSUB_LOOKUP_TYPE_ALTERNATE, GSUB_LOOKUP_TYPE_CHAINED_CONTEXT, GSUB_LOOKUP_TYPE_CONTEXT,
+    GSUB_LOOKUP_TYPE_EXTENSION, GSUB_LOOKUP_TYPE_LIGATURE, GSUB_LOOKUP_TYPE_MULTIPLE,
+    GSUB_LOOKUP_TYPE_REVERSE_CHAINED_SINGLE, GSUB_LOOKUP_TYPE_SINGLE,
 };
 pub use crate::tables::layout::{
     Feature, FeatureList, FeatureListIter, LangSys, Lookup, LookupFlag, LookupList, LookupListIter,